refactor(datatool): modularization
