+++ bin/datatool/src/args.rs
@@ -1,0 +1,185 @@
+use std::path::PathBuf;
+
+use argh::FromArgs;
+use bitcoin::Network;
+use rand::rngs::OsRng;
+
+/// Args.
+#[derive(FromArgs)]
+pub(crate) struct Args {
+    #[argh(option, description = "network name [signet, regtest]", short = 'b')]
+    pub(crate) bitcoin_network: Option<String>,
+
+    #[argh(
+        option,
+        description = "data directory (unused) (default cwd)",
+        short = 'd'
+    )]
+    pub(crate) datadir: Option<PathBuf>,
+
+    #[argh(subcommand)]
+    pub(crate) subc: SubcommandGen,
+}
+
+#[derive(FromArgs, PartialEq, Debug)]
+#[argh(subcommand)]
+pub(crate) enum SubcommandGen {
+    Seed(SubcSeed),
+    SeqPubkey(SubcSeqPubkey),
+    SeqPrivkey(SubcSeqPrivkey),
+    OpXpub(SubcOpXpub),
+    Params(SubcParams),
+}
+
+#[derive(FromArgs, PartialEq, Debug)]
+#[argh(
+    subcommand,
+    name = "genseed",
+    description = "generates a xpriv and writes it to a file"
+)]
+pub(crate) struct SubcSeed {
+    #[argh(positional, description = "output path")]
+    pub(crate) path: PathBuf,
+
+    #[argh(switch, description = "force overwrite", short = 'f')]
+    pub(crate) force: bool,
+}
+
+/// Generate the sequencer pubkey to pass around.
+#[derive(FromArgs, PartialEq, Debug)]
+#[argh(
+    subcommand,
+    name = "genseqpubkey",
+    description = "generates a sequencer pubkey from seed"
+)]
+pub(crate) struct SubcSeqPubkey {
+    #[argh(option, description = "reads key from specified file", short = 'f')]
+    pub(crate) key_file: Option<PathBuf>,
+
+    #[argh(
+        switch,
+        description = "reads key from envvar STRATA_SEQ_KEY",
+        short = 'E'
+    )]
+    pub(crate) key_from_env: bool,
+}
+
+/// Generate the sequencer pubkey to pass around.
+#[derive(FromArgs, PartialEq, Debug)]
+#[argh(
+    subcommand,
+    name = "genseqprivkey",
+    description = "generates a sequencer privkey from seed"
+)]
+pub(crate) struct SubcSeqPrivkey {
+    #[argh(option, description = "reads key from specified file", short = 'f')]
+    pub(crate) key_file: Option<PathBuf>,
+
+    #[argh(
+        switch,
+        description = "reads key from envvar STRATA_SEQ_KEY",
+        short = 'E'
+    )]
+    pub(crate) key_from_env: bool,
+}
+
+/// Generate operator xpub to pass around.
+#[derive(FromArgs, PartialEq, Debug)]
+#[argh(
+    subcommand,
+    name = "genopxpub",
+    description = "generates an operator xpub from seed"
+)]
+pub(crate) struct SubcOpXpub {
+    #[argh(option, description = "reads key from specified file", short = 'f')]
+    pub(crate) key_file: Option<PathBuf>,
+
+    #[argh(
+        switch,
+        description = "reads key from envvar STRATA_OP_KEY",
+        short = 'E'
+    )]
+    pub(crate) key_from_env: bool,
+}
+
+/// Generate a network's param file from inputs.
+#[derive(FromArgs, PartialEq, Debug)]
+#[argh(
+    subcommand,
+    name = "genparams",
+    description = "generates network params from inputs"
+)]
+pub(crate) struct SubcParams {
+    #[argh(
+        option,
+        description = "output file path .json (default stdout)",
+        short = 'o'
+    )]
+    pub(crate) output: Option<PathBuf>,
+
+    #[argh(
+        option,
+        description = "network name, used for magics (default random)",
+        short = 'n'
+    )]
+    pub(crate) name: Option<String>,
+
+    #[argh(
+        option,
+        description = "sequencer pubkey (default unchecked)",
+        short = 's'
+    )]
+    pub(crate) seqkey: Option<String>,
+
+    #[argh(
+        option,
+        description = "add a bridge operator key (must be at least one, appended after file keys)",
+        short = 'b'
+    )]
+    pub(crate) opkey: Vec<String>,
+
+    #[argh(
+        option,
+        description = "read bridge operator keys by line from file",
+        short = 'B'
+    )]
+    pub(crate) opkeys: Option<PathBuf>,
+
+    #[argh(option, description = "deposit amount in sats (default \"10 BTC\")")]
+    pub(crate) deposit_sats: Option<String>,
+
+    #[argh(
+        option,
+        description = "genesis trigger height (default 100)",
+        short = 'g'
+    )]
+    pub(crate) genesis_trigger_height: Option<u64>,
+
+    #[argh(
+        option,
+        description = "block time in seconds (default 15)",
+        short = 't'
+    )]
+    pub(crate) block_time: Option<u64>,
+
+    #[argh(option, description = "epoch duration in slots (default 64)")]
+    pub(crate) epoch_slots: Option<u32>,
+
+    #[argh(
+        option,
+        description = "permit blank proofs after timeout in millis (default strict)"
+    )]
+    pub(crate) proof_timeout: Option<u32>,
+}
+
+pub(crate) struct CmdContext {
+    /// Resolved datadir for the network.
+    #[allow(unused)]
+    pub(crate) datadir: PathBuf,
+
+    /// The Bitcoin network we're building on top of.
+    pub(crate) bitcoin_network: Network,
+
+    /// Shared RNG, just [`OsRng`] for now.
+    pub(crate) rng: OsRng,
+}
